+++ htdocs/custom/bankconnect/class/Candidate.php
@@ -5,7 +5,7 @@ class Candidate
     /** @var int|string */
     public $id = 0;
 
-    /** @var string supplier_invoice|customer_invoice|expense */
+    /** @var string supplier_invoice|customer_invoice|supplier_payment|customer_payment|various_payment */
     public string $type = 'supplier_invoice';
 
     public string $ref = '';
@@ -13,4 +13,5 @@ class Candidate
     public float $remaining = 0.0;
     public string $date = '';
     public string $thirdparty = '';
+    public string $currency = 'DKK';
 }